move files to a benchmark-folder
