+++ README.markdown
@@ -25,8 +25,11 @@ It's not designed for
 ## Usage
 With Docker-compose
 
-    docker-compose build; docker-compose down; docker-compose run mmq-load-tests -- status.sh
-    docker-compose build; docker-compose down; docker-compose run mmq-load-tests -- post_message.sh
+    docker-compose build; docker-compose up;
+    docker-compose run mmq-load-tests -- status.sh
+  
+Or, 
+    docker-compose run mmq-load-tests -- post_message.sh
 
 Enter load-tests container
