+++ benchmark/service-nodejs/src/logger.js
