--- load-tests/post-json2.lua
@@ -1,3 +1,0 @@
-wrk.method = "POST"
-wrk.body   = "{\"id\":100}"
-wrk.headers["Content-Type"] = "application/json"
