--- service-kestrel/Service-Kestrel/Service-Kestrel/ApplicationLogging.cs
@@ -1,12 +1,0 @@
-﻿using Microsoft.Extensions.Logging;
-
-namespace Service_Kestrel
-{
-	internal static class ApplicationLogging
-	{
-		internal static ILoggerFactory LoggerFactory { get; set; } // = new LoggerFactory();
-		internal static ILogger CreateLogger<T>() => LoggerFactory.CreateLogger<T>();
-		internal static ILogger CreateLogger(string categoryName) => LoggerFactory.CreateLogger(categoryName);
-
-	}
-}
