+++ service-kestrel/Service-Kestrel/Service-Kestrel/Exceptions/LikelyThreadStarvationException.cs
@@ -1,0 +1,12 @@
+﻿using Microsoft.AspNetCore.Server.Kestrel.Core;
+using System;
+
+namespace Service_Kestrel.Exceptions
+{
+	public class SuspectedThreadStarvationException : Exception
+	{
+		public SuspectedThreadStarvationException(BadHttpRequestException exception) : base(exception.Message, exception)
+		{
+		}
+	}
+}
