+++ service-kestrel/Service-Kestrel/Service-Kestrel/ServiceKestrelOptions.cs
@@ -1,0 +1,7 @@
+﻿namespace Service_Kestrel
+{
+	public class ServiceKestrelOptions
+	{
+		public string FasterDevice { get; set; }
+	}
+}
